feat: migrate scientific-method server to code mode architecture
- Extract core business logic into standalone ScientificMethodCore class
- Implement typed Code Mode API (ScientificMethodCodeMode) for programmatic use
- Create MCP adapter layer that wraps the Code Mode API
- Reorganize project structure into core/, codemode/, and mcp/ directories
- Update build configuration to use tsconfig.build.json for production builds
- Add comprehensive test suite for the new Code Mode API
- Remove old MCP-specific tests in favor of direct API testing
