+++ .trae/documents/plan_20260212_230858.md
@@ -1,0 +1,72 @@
+This plan outlines the migration of the `scientific-method` MCP server to the Code Mode architecture, separating concerns into Core logic, a programmable Code Mode API, and an MCP adapter layer.
+
+### 1. Preparation & Structure Setup
+
+* Create the standard directory structure:
+
+  * `src/core/`: For pure business logic and types.
+
+  * `src/codemode/`: For the public TypeScript API.
+
+  * `src/mcp/`: For MCP tool definitions and server adaptation.
+
+* Update `tsconfig.json` to support the new structure and ensure proper build output.
+
+### 2. Core Layer Extraction (`src/core/`)
+
+* **Types (`src/core/types.ts`)**: Extract all interfaces (`ScientificInquiryData`, `HypothesisData`, `ExperimentData`, etc.) from `src/index.ts`.
+
+* **Logic (`src/core/ScientificMethod.ts`)**:
+
+  * Move the `ScientificMethodServer` class logic here.
+
+  * Refactor `processScientificInquiry` to return typed objects (`ScientificInquiryData`) instead of MCP-formatted JSON strings.
+
+  * Remove MCP-specific dependencies (like `@modelcontextprotocol/sdk`) from this layer.
+
+  * Ensure state management (`inquiryHistory`, `hypothesisRegistry`) is preserved within the class instance.
+
+### 3. Code Mode Layer Implementation (`src/codemode/`)
+
+* **API (`src/codemode/index.ts`)**:
+
+  * Create a `ScientificMethodCodeMode` class that wraps the Core `ScientificMethod` class.
+
+  * Expose a clean, typed method `processInquiry(input: ScientificInquiryInput): Promise<ScientificInquiryData>`.
+
+  * Ensure this layer provides a strictly typed interface suitable for programmatic use by LLMs.
+
+### 4. MCP Adapter Layer Implementation (`src/mcp/`)
+
+* **Tools (`src/mcp/tools.ts`)**: Extract the `SCIENTIFIC_METHOD_TOOL` definition into a dedicated file.
+
+* **Server (`src/mcp/index.ts`)**:
+
+  * Create the MCP server setup that instantiates `ScientificMethodCodeMode`.
+
+  * Implement the `CallToolRequest` handler to call the Code Mode API and format the result into the MCP-expected structure (JSON string inside a content block).
+
+### 5. Entry Point & Configuration
+
+* **Entry Point (`src/index.ts`)**:
+
+  * Update to export the Code Mode API as the default export.
+
+  * Add conditional logic to run the MCP server if executed directly (e.g., `if (import.meta.main) ...`).
+
+* **Package Configuration**:
+
+  * Update `package.json` `exports` and `types` to point to the new Code Mode API.
+
+  * Ensure build scripts (`bun build`) correctly bundle the application.
+
+### 6. Testing & Verification
+
+* **Tests**:
+
+  * Refactor `src/index.test.ts` to test the Code Mode API (`ScientificMethodCodeMode`) directly instead of the MCP server.
+
+  * Verify that the logic correctly handles all stages of the scientific method (observation, hypothesis, experiment, etc.).
+
+* **Validation**: Ensure backward compatibility by verifying the MCP server still responds correctly to tool calls.
+
