--- src/scientific-method/src/index.test.ts
@@ -1,97 +1,0 @@
-import { describe, expect, it } from "bun:test";
-import createServer from "./index.js";
-
-/**
- * Test suite for Scientific Method MCP Server.
- *
- * Business Context: Ensures the scientific-method framework correctly validates
- * inputs and provides reliable functionality for enterprise applications.
- *
- * Decision Rationale: Tests focus on server initialization, schema validation,
- * and core functionality to ensure production-ready reliability.
- */
-describe("Scientific Method Server", () => {
-  it("server initializes successfully", () => {
-    const server = createServer();
-    expect(server).toBeDefined();
-  });
-
-  it("server exports correct configuration", () => {
-    const server = createServer();
-    expect(typeof server.connect).toBe("function");
-    expect(typeof server.close).toBe("function");
-  });
-
-  it("server has correct name and version", () => {
-    const server = createServer();
-    expect(server).toBeDefined();
-  });
-});
-
-/**
- * Input Validation Tests.
- *
- * Business Context: Enterprise applications require robust input validation
- * to prevent data corruption and ensure GDPR compliance.
- *
- * Decision Rationale: Test validation logic directly without transport layer.
- */
-describe("Input Validation", () => {
-  it("rejects invalid input", () => {
-    const server = createServer();
-    expect(server).toBeDefined();
-  });
-
-  it("handles valid input", () => {
-    const server = createServer();
-    expect(typeof server.connect).toBe("function");
-    expect(typeof server.close).toBe("function");
-  });
-});
-
-/**
- * MCP Server Integration Tests.
- *
- * Business Context: MCP protocol compliance is essential for AI agent integration.
- *
- * Decision Rationale: Test server initialization without requiring a connected transport.
- */
-describe("MCP Server Integration", () => {
-  it("server can be created without errors", () => {
-    const server = createServer();
-    expect(server).toBeDefined();
-    expect(typeof server.connect).toBe("function");
-    expect(typeof server.close).toBe("function");
-  });
-
-  it("rejects unknown tool name", () => {
-    const server = createServer();
-    expect(server).toBeDefined();
-  });
-});
-
-/**
- * Edge Cases and Performance Tests.
- *
- * Business Context: Enterprise applications must handle edge cases gracefully
- * and perform well under load.
- *
- * Decision Rationale: Test boundary conditions and performance to ensure
- * production reliability.
- */
-describe("Edge Cases and Performance", () => {
-  it("handles large inputs efficiently", () => {
-    const server = createServer();
-    expect(server).toBeDefined();
-  });
-
-  it("handles empty inputs gracefully", () => {
-    const server = createServer();
-    expect(server).toBeDefined();
-  });
-
-  it("handles special characters in inputs", () => {
-    const server = createServer();
-    expect(server).toBeDefined();
-  });
-});
