+++ src/scientific-method/tsconfig.build.json
@@ -1,0 +1,4 @@
+{
+  "extends": "./tsconfig.json",
+  "exclude": ["node_modules", "dist", "tests", "**/*.test.ts", "src/index.test.ts"]
+}
